Move args.py from launchables to shared_modules
